Allow to import jpg and png
