+++ .idea/dictionaries/rapha.xml
@@ -10,6 +10,7 @@
       <w>fgsm</w>
       <w>imagestack</w>
       <w>imagestacks</w>
+      <w>imgstack</w>
       <w>logdir</w>
       <w>matplot</w>
       <w>matplotlib</w>
@@ -18,6 +19,7 @@
       <w>preprocessing</w>
       <w>pretrain</w>
       <w>pyplot</w>
+      <w>segmentations</w>
       <w>tensorboard</w>
       <w>testset</w>
       <w>visboard</w>
