+++ .idea/runConfigurations/ds_add_image_folder.xml
@@ -7,13 +7,13 @@
       <env name="PYTHONUNBUFFERED" value="1" />
     </envs>
     <option name="SDK_HOME" value="C:\Users\rapha\Desktop\ai-trainer\envs\python.exe" />
-    <option name="WORKING_DIRECTORY" value="$PROJECT_DIR$/../tutorials/data" />
+    <option name="WORKING_DIRECTORY" value="$PROJECT_DIR$/../tutorials/data/test_dataset" />
     <option name="IS_MODULE_SDK" value="false" />
     <option name="ADD_CONTENT_ROOTS" value="true" />
     <option name="ADD_SOURCE_ROOTS" value="true" />
     <EXTENSION ID="PythonCoverageRunConfigurationExtension" runner="coverage.py" />
     <option name="SCRIPT_NAME" value="trainer.tools.cli" />
-    <option name="PARAMETERS" value="ds add-image-folder -p test_dataset" />
+    <option name="PARAMETERS" value="ds add-image-folder" />
     <option name="SHOW_COMMAND_LINE" value="true" />
     <option name="EMULATE_TERMINAL" value="false" />
     <option name="MODULE_MODE" value="true" />
