+++ .idea/runConfigurations/ds_annotate.xml
@@ -7,7 +7,7 @@
       <env name="PYTHONUNBUFFERED" value="1" />
     </envs>
     <option name="SDK_HOME" value="C:\Users\rapha\Desktop\ai-trainer\envs\python.exe" />
-    <option name="WORKING_DIRECTORY" value="$PROJECT_DIR$/../tutorials/data/crucial_ligament_diagnosis" />
+    <option name="WORKING_DIRECTORY" value="$PROJECT_DIR$/../tutorials/data/test_dataset" />
     <option name="IS_MODULE_SDK" value="false" />
     <option name="ADD_CONTENT_ROOTS" value="true" />
     <option name="ADD_SOURCE_ROOTS" value="true" />
