+++ trainer/__init__.py
@@ -1,1 +1,1 @@
-VERSION = "0.0.7"
+VERSION = "0.0.8"
