+++ trainer/tools/cli.py
@@ -10,8 +10,9 @@ import click
 
 from trainer.lib import standalone_foldergrab
 from trainer.lib.JsonClass import dir_is_json_class
-from trainer.ml.data_model import Dataset
 from trainer.tools.AnnotationGui import AnnotationGui, run_window
+import trainer.ml as ml
+import trainer.lib as lib
 
 
 @click.group()
@@ -48,7 +49,7 @@ def dataset_init(parent_path, name):
         if os.path.isdir(p):
             click.echo(f"Dirname: {os.path.basename(p)}")
     if click.confirm(f"The dataset {name} will be created in {parent_path}"):
-        d = Dataset.build_new(name, parent_path)
+        d = ml.Dataset.build_new(name, parent_path)
         d.to_disk(parent_path)
     click.echo(f"For working with the dataset {name}, please switch into the directory")
 
@@ -70,7 +71,7 @@ def dataset_annotate(dataset_path: str, subject_name: str):
     """
     if not subject_name:
         # Subject name needs to be picked
-        d = Dataset.from_disk(dataset_path)
+        d = ml.Dataset.from_disk(dataset_path)
         subject_name = d.get_subject_name_list()[0]  # Just pick the first subject
     run_window(AnnotationGui, os.path.join(dataset_path, subject_name), dataset_path)
 
@@ -98,7 +99,7 @@ def dataset_train(dataset_path: str):
     """
     if not dir_is_json_class(dataset_path):
         raise Exception("The given directory is not a valid Dataset")
-    d = Dataset.from_disk(dataset_path)
+    d = ml.Dataset.from_disk(dataset_path)
 
     seg_structs = d.compute_segmentation_structures()
 
@@ -112,9 +113,9 @@ def dataset_train(dataset_path: str):
 @click.option('--dataset-path', '-p', default=os.getcwd)
 @click.option('--subject-name', '-s', default='')
 def dataset_visualize(dataset_path: str, subject_name: str):
-    ds = Dataset.from_disk(dataset_path)
+    d = ml.Dataset.from_disk(dataset_path)
     if subject_name:
-        s = ds.get_subject_by_name(subject_name)
+        s = d.get_subject_by_name(subject_name)
         s.matplot_imagestacks()
     else:
         # TODO Create a html folder with a simple static webview
@@ -127,16 +128,16 @@ def dataset_visualize(dataset_path: str, subject_name: str):
 @click.option('--folder-path', '-ip', default='')
 @click.option('--structure-tpl', '-st', default='')
 def dataset_add_image_folder(dataset_path: str, folder_path: str, structure_tpl: str):
-    ds = Dataset.from_disk(dataset_path)
+    d = ml.Dataset.from_disk(dataset_path)
     if not folder_path:
         folder_path, inputs_dict = standalone_foldergrab(
             folder_not_file=True,
             title='Pick Image folder',
-            optional_choices=[('Structure Template', 'str_tpl', ds.get_structure_template_names())]
+            optional_choices=[('Structure Template', 'str_tpl', d.get_structure_template_names())]
         )
         structure_tpl = inputs_dict['str_tpl']
-    seg_structs = ds.get_structure_template_by_name(structure_tpl)
-    ds.add_image_folder(folder_path, structures=seg_structs)
+    seg_structs = d.get_structure_template_by_name(structure_tpl)
+    lib.add_image_folder(d, folder_path, structures=seg_structs)
 
 
 if __name__ == '__main__':
